move util functions to separate package
