--- tests/knownValues.go
@@ -1,6 +1,0 @@
-package tests
-
-const (
-	// Tests expect a ledger device initialized to the following mnemonic
-	TestMnemonic = "equip will roof matter pink blind book anxiety banner elbow sun young"
-)
