+++ testsUtil/message.go
@@ -1,0 +1,55 @@
+package testUtil
+
+import sdk "github.com/pokt-network/posmint/types"
+
+
+const (
+	routeMsgCounter  = "msgCounter"
+	routeMsgCounter2 = "msgCounter2"
+)
+
+type msgCounter struct {
+	Counter       int64
+	FailOnHandler bool
+}
+
+// Implements Msg
+func (msg msgCounter) Route() string                { return routeMsgCounter }
+func (msg msgCounter) Type() string                 { return "counter1" }
+func (msg msgCounter) GetSignBytes() []byte         { return nil }
+func (msg msgCounter) GetSigners() []sdk.AccAddress { return nil }
+func (msg msgCounter) ValidateBasic() sdk.Error {
+	if msg.Counter >= 0 {
+		return nil
+	}
+	return sdk.ErrInvalidSequence("counter should be a non-negative integer.")
+}
+
+type msgCounter2 struct {
+	Counter int64
+}
+
+// Implements Msg
+func (msg msgCounter2) Route() string                { return routeMsgCounter2 }
+func (msg msgCounter2) Type() string                 { return "counter2" }
+func (msg msgCounter2) GetSignBytes() []byte         { return nil }
+func (msg msgCounter2) GetSigners() []sdk.AccAddress { return nil }
+func (msg msgCounter2) ValidateBasic() sdk.Error {
+	if msg.Counter >= 0 {
+		return nil
+	}
+	return sdk.ErrInvalidSequence("counter should be a non-negative integer.")
+}
+
+type msgNoRoute struct {
+	msgCounter
+}
+
+func (tx msgNoRoute) Route() string { return "noroute" }
+
+// a msg we dont know how to decode
+type msgNoDecode struct {
+	msgCounter
+}
+
+func (tx msgNoDecode) Route() string { return routeMsgCounter }
